+++ ewoms/disc/vcfv/vcfvdiscretization.hh
@@ -134,7 +134,7 @@ class VcfvDiscretization : public FvBaseDiscretization<TypeTag>
 
 
 public:
-    VcfvDiscretization(Simulator &simulator)
+    VcfvDiscretization(Simulator& simulator)
         : ParentType(simulator)
     { }
 
@@ -148,13 +148,13 @@ public:
      * \brief Returns the number of global degrees of freedom (DOFs) due to the grid
      */
     size_t numGridDof() const
-    { return this->gridView_.size(/*codim=*/dim); }
+    { return static_cast<size_t>(this->gridView_.size(/*codim=*/dim)); }
 
     /*!
      * \brief Mapper to convert the Dune entities of the
      *        discretization's degrees of freedoms are to indices.
      */
-    const DofMapper &dofMapper() const
+    const DofMapper& dofMapper() const
     { return this->simulator_.problem().vertexMapper(); }
 
     /*!
@@ -165,7 +165,7 @@ public:
      * \param res The serializer object
      */
     template <class Restarter>
-    void serialize(Restarter &res)
+    void serialize(Restarter& res)
     { res.template serializeEntities</*codim=*/dim>(asImp_(), this->gridView_); }
 
     /*!
@@ -176,16 +176,16 @@ public:
      * \param res The serializer object
      */
     template <class Restarter>
-    void deserialize(Restarter &res)
+    void deserialize(Restarter& res)
     {
         res.template deserializeEntities</*codim=*/dim>(asImp_(), this->gridView_);
         this->solution(/*timeIdx=*/1) = this->solution(/*timeIdx=*/0);
     }
 
 private:
-    Implementation &asImp_()
+    Implementation& asImp_()
     { return *static_cast<Implementation*>(this); }
-    const Implementation &asImp_() const
+    const Implementation& asImp_() const
     { return *static_cast<const Implementation*>(this); }
 };
 } // namespace Ewoms
